UI Card detail : Externalize card footer text to another component
