+++ ui/main/src/app/modules/cards/components/card-footer-text/card-footer-text.component.html
@@ -1,0 +1,27 @@
+<!-- Copyright (c) 2022, RTE (http://www.rte-france.com)              -->
+<!-- See AUTHORS.txt                                                       -->
+<!-- This Source Code Form is subject to the terms of the Mozilla Public   -->
+<!-- License, v. 2.0. If a copy of the MPL was not distributed with this   -->
+<!-- file, You can obtain one at http://mozilla.org/MPL/2.0/.              -->
+<!-- SPDX-License-Identifier: MPL-2.0                                      -->
+
+
+<div class="opfab-card-received-footer" style="text-align: left;white-space: pre-line;">
+      <div>
+            <span translate="feed.received"></span> {{formattedPublishDate}}
+            <span translate="feed.at"></span> {{formattedPublishTime}}
+            <span *ngIf="!!fromEntityOrRepresentative && fromEntityOrRepresentative.length > 0"
+                  translate="feed.from"></span> {{fromEntityOrRepresentative}}
+      </div>
+      <div id="opfab-card-details-address-to" *ngIf="listEntitiesAcknowledged.length > 1">
+            <span translate="feed.addressedTo"></span>
+            <span *ngFor="let entity of listEntitiesAcknowledged; let isLast = last"> {{entity.entityName}} <em
+                        *ngIf="entity.acknowledged" style="color: #0082FF;" class="fa fa-check"></em><span
+                        *ngIf="!isLast" style="margin-right: 5px;">, </span></span>
+      </div>
+      <div *ngIf="!!this.lastResponse">
+            <span translate="feed.lastResponse"></span> {{formatDate(lastResponse.publishDate)}}
+            <span translate="feed.at"></span> {{formatTime(lastResponse.publishDate)}}
+            <span translate="feed.from"></span> {{getResponsePublisher(lastResponse)}}
+      </div>
+</div>
