+++ ui/main/src/app/modules/cards/components/card-footer-text/card-footer-text.component.scss
@@ -1,0 +1,16 @@
+
+/* Copyright (c) 2022, RTE (http://www.rte-france.com)
+ * See AUTHORS.txt
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ * SPDX-License-Identifier: MPL-2.0
+ * This file is part of the OperatorFabric project.
+ */
+
+.opfab-card-received-footer {
+    float: left;
+    font-size: 14px;
+    font-weight: 600;
+    margin: 10px;
+}
