+++ ui/main/src/app/modules/cards/components/detail/detail.component.scss
@@ -134,12 +134,6 @@
 .flex-container {
     display: flex;
 }
-.opfab-card-received-footer {
-    float: left;
-    font-size: 14px;
-    font-weight: 600;
-    margin: 10px;
-}
